Refactor routes into Api modules
