+++ internal/app/server.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io"
 	"log"
 	"net/http"
 	"os"
@@ -12,48 +11,27 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/koubae/game-hangar/internal/app/api"
+	"github.com/koubae/game-hangar/internal/app/settings"
 	"github.com/koubae/game-hangar/pkg/common"
-	"github.com/koubae/game-hangar/pkg/middleware"
-	"github.com/rs/cors"
 	"go.uber.org/zap"
 )
 
 func RunServer() {
 	loggerTmp, _ := common.CreateLogger(common.LogLevelInfo, "")
-	config := NewConfig(loggerTmp)
+	config := settings.NewConfig(loggerTmp)
 
 	logger, logCloser := common.CreateLogger(config.LogLevel, config.LogFilePath)
 	defer logCloser(loggerTmp, logger)
 
-	mux := http.NewServeMux()
-	mux.HandleFunc(
-		"GET /{$}", func(w http.ResponseWriter, req *http.Request) {
-			w.WriteHeader(http.StatusOK)
-			_, err := io.WriteString(w, fmt.Sprintf("Welcome to %s", config.GetFullName()))
-			if err != nil {
-				return
-			}
-		},
-	)
-
-	handler := cors.New(
-		cors.Options{
-			AllowedOrigins:   config.CORSConfig.AllowedOrigins,
-			AllowedMethods:   config.CORSConfig.AllowedMethods,
-			AllowedHeaders:   config.CORSConfig.AllowedHeaders,
-			AllowCredentials: config.CORSConfig.AllowCredentials,
-		},
-	).Handler(mux)
-	handler = middleware.AccessLogger(logger, handler)
-	handler = middleware.RecoveryMiddleware(logger, handler)
-
+	routerHandler := api.Router(logger)
 	srv := http.Server{
 		Addr:           config.GetAppURL(),
 		ReadTimeout:    time.Duration(config.ServerReadTimeout) * time.Second,
 		WriteTimeout:   time.Duration(config.ServerWriteTimeout) * time.Second,
 		IdleTimeout:    time.Duration(config.ServerIdleTimeout) * time.Second,
 		MaxHeaderBytes: config.ServerMaxHeaderBytes,
-		Handler:        handler,
+		Handler:        *routerHandler,
 	}
 
 	go func() {
