+++ internal/app/settings/config.go
@@ -1,4 +1,4 @@
-package app
+package settings
 
 import (
 	"fmt"
